Moved double clicks to dbpanel, now works for customers too
